Changed data output system
Removed the ability to output data to ascii files. Instead, all
output data is stored in a ROOT TTree and is written to a root
file. There are two possible trees included in this file. "VIKAR"
is the standard output tree and contains data for ejectile and
recoil particles. Entries are only written to this tree if an
ejectile is detected in one of the detectors. "DEBUG" is an
optional tree that contains generic debug information. Entries
are written to DEBUG every iteration regardless of whether or not
a detector saw an event.