V1.11
Included beamspot effects to the simulation. Added
cylindrical and conical beam shapes (via a new parameter
in the input file named 'beam angular divergence'. For
both beam types, the program will calculate a random
vector within the beam that intersects with the target.

The target now includes a planar object to make it easier
to calculate beam particle intersections with the target.

Planar now supports tracing of rays which do not originate
at the origin. This makes it possible to trace rays from
a target of non-negligible size.

The Target class now handles angular straggling instead
of using separate functions from vikar_core.

RangeTable now stores particle ranges in meters instead of
mg/cm^2. This was necessary since all of the geometry is
now measured in meters.

TO DO:
 Fix Target::AngleStraggling() and transform() to get them
 to accept cartesian vectors instead of polar angles.

 The reaction vectors should now be considered rays which
 originate at the reaction position inside the target. They
 no longer originate from the origin.