V1.13
Introduced new config file option to allow Monte Carlo
test of detector setups. The geometric efficiency of
the setup is output and the user is asked if they
wish to continue with the simulation.

Introduced a new Matrix3 class for the handling of
rotational matrices and vector transformations.

For each beam particle event, a rotational matrix is
calculated and is used to transform both the ejectile
and the recoil vectors from the beam particle frame
to the lab frame. Angular straggling transformations
are still turned off, however.

The program now tracks the number of particles which
are stopped in the target and will warn the user if
the number of stopped particles becomes large (10,000).