V1.13d
Included a new data type called 'ReactionObject' to
structures.cpp. This data type is used to store
information about the reaction. Namely, the interaction
point inside the target, the trajectory of the beam
particle, and the energy of the beam particle prior
to the reaction.

Added a new option in the config file to turn this
reaction output on or off.