V1.13f
Turned off detector efficency considerations (even
when perfect detectors are turned off). This was
done in order to implement energy losses for the
ejectile and recoil particles.

Vikar now calculates range tables for beam in
target, and ejectiles and recoils in all loaded
materials. When a detector is hit by an ejectile
or a recoil, vikar checks if the detector has an
associated material. If it does, the particle will
lose an amount of energy dependent on the distance
it travels through the detector. If it doesn't, the
particle will move a random distance through the
detector and will deposit a random amount of energy.

Energy losses are now tracked. The amount of energy
left in a detector (QDC) is subtracted from the
original particle energy before checking for
detections in other detectors.