V1.15
Moved section of vikar.cpp where detectors are loaded
to before initialization of materials. This was done to
avoid needlessly calculating range tables if the user
only wanted to run a Monte Carlo detector setup test.

Removed "supply detector setup file" option from input
file. The option to not supply a detector setup file
was not implemented in the code anyway.

Beamspot now represents the FWHM of a gaussian beam, not
the absolute diameter of a perfectly cylindrical beam.

TestDetSetup now uses beam FWHM and target angle if they
are non-zero in the input file. If the specified FWHM is
zero, an isotropic source is used.

Vikar will now ask the user if they would like to
continue the simulation after a Monte Carlo test is
performed.