V1.15b
Fixed a programming error where detectors were being
initialized with an invalid material. This resulted in
detectors not using energy loss tables and instead
stopping the particle a random depth into the detector.
This caused the output energy (qdc) to be smeared along
the entire energy spectrum.

Added a new material type 'C12'.