V1.15c
Added some documentation to detectors.cpp and
detectors.h

Added a new overload of Kindeux::FillVars and
added default parameters to the original method
in order to fix tools/Kindist.cpp. The default
parameters allow the user to manually select a
recoil state, +/- solution, and/or a center of
mass angle.

Added a new material 'CH2'.