V1.15d
Removed target density parameter from Kindeux class.
This parameter is now passed to Kindeux::SetDist
instead of being passed to the Kindeux::Initialize
method.

Kindeux::FillVars now takes an additional &double
parameter which is used to return the center of
mass angle.

ReactionObject now also stores the center of
mass angle of the reaction.

Added a new TestDetSetup method which outputs an
ascii histogram instead of a root file. This file
contains geometric efficiencies for center of
mass angles. This new method must be passed a
pointer to a Kindeux object.

Moved initialization of Kindeux object to just
before monte carlo test in vikar.cpp so that
the Vikar Kindeux object may be passed to
TestDetSetup.