V1.15e
Added a new overload for Cart2Sphere and
Sphere2Cart which takes an input vector and
returns three doubles.

Changed the way TestDetSetup calculates the
center of mass angle of an event. It no longer
uses the center of mass angle returned from
Kindeux::FillVars. Instead, it calculates the
angle using the ejectile and recoil vectors.