V1.17
Added ability to perform Rutherford scattering. If
"supply angular distributions" is set to true and
"RUTHERFORD" is passed as the distribution filename,
Vikar will calculate the Rutherford scattering
cross section and will use that distribution for
generating ejectile particles.

Added Kindex::SetRutherford method for setting the
use of a Rutherford scattering distribution.

Added an overload of AngularDist::Initialize which
allows the supply of angles and cross section
values from an external source.