V1.18
Added beam type option to the input file. User may
specify either gaussian beam (0) or a conical beam
(1). For a gaussian beam, the "beamspot" parameter
represents the beam size FWHM at the target (z=0 m).
For a conical beam, the beamspot represents the
diameter of the beamspot at the target (z=0 m). A
cylindrical beam may be produced by setting the
beam angular divergence to zero.

Changed the way the beam angular divergence is
measured. It is now 90 - theta, for increased
clarity. A perfectly cylindrical beam, for example,
is generated by using a divergence of 0. Previously,
this beam would be generated using a divergence of 90.

Removed upstream and downstream focal point
considerations from the main routine. This
consideration is now handled within vikar_core
using "RandomCircle" and "RandomGauss" routines.
These functions select upstream or downstream focus
based on the sign of the offset_ parameter.