V1.19b
Added support for "NONE" targets. This will specify
that energy loss calculations are not to be done for
beam particles and reaction products in the target.
Deactivating target energy loss also deactivates
angular straggling in the target.

Range tables are now only calculated for materials
which are called for by the detector file.