V1.22
ReadDetFile now takes a vector of Planar* by ref
instead of a pointer to a Planar array. This was
done in order to allow the use of classes which
are derived from Planar.

Added new Planar derived classes Elliptical, Polygonal,
and Annular. These are accessible by setting the
subtype to "ellipse", "polygon", or "annular".

Added new class RegularPolygon which is used to
determine whether a point in 2d space is inside of
a n-sided regular polygon.

Added "Ray" and "Line" classes to vikar_core. The
Ray class is used to define a ray in 2d space. The
Line class is used to define a line segment bounded
by two points in 2d space.