Added new geometry classes
Added new geometry classes Planar, Cylindrical, and
Spherical. Planar should contain the same level of
functionality as the old Planar class.

Added "cylinder" and "sphere" as detector subtypes.
The radius of the cylinder is equal to half of its
width while the radius of the sphere is equal to
half of its length.