Added qt "renderer" executable
Added psi angle to SetRotationMatrixSphere method
of Matrix3.

Added CosAngle method to Vector3. This method returns
the cosine of the angle between two vectors.

Fixed a coding error in testSetup.cpp. The old Planar
class was still being used instead of Primitive.

Added qt files for setup viewing executable named
"renderer". Renderer is in the tools directory but
is only built if "make renderer" is called
explicitly. This is done to avoid errors cluttering up
the terminal on machines which do not have qt installed.