V1.25
Failed interpolation by RangeTable::_interpolate now
returns zero in the event of the input value being
less than that of the first entry in the table.

RangeTable::GetNewE sets total distance to the input
penetration distance if the range > penetration (i.e.
the particle is energetic enough to exit the material).

Set the minimum range table energy from 0.1 MeV to the
KE of the particle with a velocity of 0.02c for the
ejectile and recoil range tables.

Correctly handle particles originating inside of a
detector i.e. a detector at the origin.

Root output lab position (_hitX, _hitY, etc) now
correctly takes detector penetration depth into
account.

Root output QDC value is now the energy deposited
in the detector by the particle.